Enabled more warning flags (-Wextra -Wno-unused-parameter -Wno-missing-field-initializers) in OS X build, and fixed several hundred resulting warnings (mostly by using more appropriate types). Bumped version number to 1.69.1. Fixed unable-to-buy-in-shipyard bug. Attempt to fix dockEscorts crash bugs (needs testing). Sundry cleanup, especially of AI and GuiDisplayGen; all exception throws now in standard form. Removed some dead code. Use correct key for Advanced Nav Array if no keyconfig.plist found. Changed OOCollectionExtractors to always clamp integer values to the requested type's range rather than truncating like NSNumber. Removed inconsistently-used and apparently-unneeded NSLock from AI.
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@1063 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Core/OOMesh.m
@@ -648,7 +648,7 @@ shaderBindingTarget:(id<OOWeakReferenceSupport>)target
     NSMutableArray		*lines;
     BOOL				failFlag = NO;
     NSString			*failString = @"***** ";
-    int					i, j;
+    unsigned			i, j;
 	NSMutableDictionary	*texFileName2Idx = nil;
 	
 	BOOL using_preloaded = NO;
@@ -674,7 +674,7 @@ shaderBindingTarget:(id<OOWeakReferenceSupport>)target
 		// strip out comments and commas between values
 		//
 		lines = [NSMutableArray arrayWithArray:[data componentsSeparatedByString:@"\n"]];
-		for (i = 0; i < [ lines count]; i++)
+		for (i = 0; i < [lines count]; i++)
 		{
 			NSString *line = [lines objectAtIndex:i];
 			NSArray *parts;
@@ -843,7 +843,7 @@ shaderBindingTarget:(id<OOWeakReferenceSupport>)target
 					if (!failFlag)
 					{
 						int vi;
-						for (i = 0; i < n_v; i++)
+						for (i = 0; (int)i < n_v; i++)
 						{
 							if ([scanner scanInt:&vi])
 							{
